% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_with_score}
\alias{feedback_with_score}
\title{MIQ feedback (with score)}
\usage{
feedback_with_score(dict = MIQ::MIQ_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual feedback at the end of the test.
}
\examples{
\dontrun{
MIQ_demo(feedback = feedback_with_score())}
}
