% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{MIQ_feedback_with_graph}
\alias{MIQ_feedback_with_graph}
\title{MIQ feedback (with graph)}
\usage{
MIQ_feedback_with_graph(dict = MIQ::MIQ_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual and graphical feedback at the end of the test.
}
\examples{
\dontrun{
MIQ_demo(feedback = MIQ_feedback_with_score())}
}
